spliting code in different files, moving obtener_matrices function to other file to fix circular import error
